PS-5735: Check not only if directory exist, also ensure that it's not a link

+++ build-ps/percona-server.spec
@@ -673,7 +673,7 @@ rm -rf $RBR%{_bindir}/ps_tokudb_admin
 ##############################################################################
 
 %pretrans -n Percona-Server-server%{product_suffix}
-if [ -d %{_datadir}/mysql ]; then
+if [ -d %{_datadir}/mysql ] && [ ! -L %{_datadir}/mysql ]; then
   MYCNF_PACKAGE=$(rpm -qi `rpm -qf /usr/share/mysql` | grep -m 1 Name | awk '{print $3}')
 fi
 
@@ -1190,7 +1190,7 @@ fi
 %endif
 
 %posttrans -n Percona-Server-server%{product_suffix}
-if [ ! -d %{_datadir}/mysql ]; then
+if [ ! -d %{_datadir}/mysql ] && [ ! -L %{_datadir}/mysql ]; then
     ln -s %{_datadir}/percona-server %{_datadir}/mysql
 fi
